move set start and end time to thunks in date time actions
